pyproject.toml
[tool.poetry]
name = "feature-fabrica"
version = "0.1.0"
description = ""
authors = ["Chingis Oinar <chingisoinar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
graphviz = "^0.20.3"
hydra-core = "^1.3.2"
pydantic = "^2.8.2"
easydict = "^1.13"
loguru = "^0.7.2"
numpy = "^2.1.0"
beartype = "^0.18.5"
scikit-learn = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:pydantic.*",
]
